MP-9 增加接口：根据币种id查询收款地址。 去掉“备注”字段。 修改id为long类型

+++ miner/miner-server/src/main/java/com/mei/hui/miner/entity/SysReceiveAddress.java
@@ -6,7 +6,6 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 
 /**
  * 收款地址表
@@ -20,7 +19,7 @@ public class SysReceiveAddress {
 
     private Long userId;
 
-    private Integer currencyId;
+    private Long currencyId;
 
     private String address;
